Move authn to /auth/authn and authz to /auth/authz
