+++ auth/authz/checker_gen.go
